Added a conditional compilation symbol that is only associated with a particular solution configuration.
Specifically there is now a solution configuration DebugMac with associated symbol 'MAC' which, if set, bypasses the authorization logic in RoleAuthHandler.
Potentially could be useful if relying on NTLM authentication but we sometimes want to develop on a Mac - we can effectively turn of authorization to ensure
it doesn't 'get in the way' of other development.